the foundation is laid out
