+++ client/components/Wordle/WordleBoard.js
@@ -1,60 +1,22 @@
 import React, { useState } from 'react';
-import { boardDefault, rows } from '.../utilities/wordle';
+import { boardDefault, rows } from './wordleUtilities/wordle';
 import LetterCell from './LetterCell';
 
-export const GameplayContext = React.createContext();
-
 function WordleBoard() {
-    const [board, setBoard] = useState(boardDefault);
-    const [attempts, setAttempts] = useState(0);
+    const [board, setBoard] = useState(rows);
+    // const [attempts, setAttempts] = useState(0);
 
     return (
         <div className="board">
-            <GameplayContext.Provider>
-                <div className='row'>
-                    {/* maybe this can be a map? */}
-                    <LetterCell index={index} attemptNum={value} />
-                    <LetterCell />
-                    <LetterCell />
-                    <LetterCell />
-                    <LetterCell />
+            {board.map((row) => (
+                <div className='row' key={row.id}>
+                    <LetterCell index={row.id} attemptNum={0} />
+                    <LetterCell index={row.id} attemptNum={1}/>
+                    <LetterCell index={row.id} attemptNum={2}/>
+                    <LetterCell index={row.id} attemptNum={3}/>
+                    <LetterCell index={row.id} attemptNum={4}/>
                 </div>
-                <div className='row'>
-                    <LetterCell />
-                    <LetterCell />
-                    <LetterCell />
-                    <LetterCell />
-                    <LetterCell />
-                </div>
-                <div className='row'>
-                    <LetterCell />
-                    <LetterCell />
-                    <LetterCell />
-                    <LetterCell />
-                    <LetterCell />
-                </div>
-                <div className='row'>
-                    <LetterCell />
-                    <LetterCell />
-                    <LetterCell />
-                    <LetterCell />
-                    <LetterCell />
-                </div>
-                <div className='row'>
-                    <LetterCell />
-                    <LetterCell />
-                    <LetterCell />
-                    <LetterCell />
-                    <LetterCell />
-                </div>
-                <div className='row'>
-                    <LetterCell />
-                    <LetterCell />
-                    <LetterCell />
-                    <LetterCell />
-                    <LetterCell />
-                </div>
-            </GameplayContext.Provider>
+            ))}
         </div>
     )
 }
